Move the principles to the recommendations section
Then add the agreed principle: that signals to the path are decoupled from e2e signals.

I also reworded the first principle, which I had some trouble with.

I considered adding a note that these are aspirational only, because we can't always eliminate signals easily.  For instance, inter-packet spacing is a signal that is used, but hard to suppress without some fairly ugly costs.  I didn't, but we might consider that at some point.

Finally, I added a security consideration regarding the potential for falsification.